+++ src/main/java/net/geforcemods/securitycraft/blockentities/RetinalScannerBlockEntity.java
@@ -80,7 +80,9 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 			if (state.getValue(RetinalScannerBlock.FACING) != hitResult.getDirection())
 				return false;
 
-			if (!state.getValue(RetinalScannerBlock.POWERED) && !EntityUtils.isInvisible(entity)) {
+			int signalLength = getSignalLength();
+
+			if ((!state.getValue(RetinalScannerBlock.POWERED) || signalLength == 0) && !EntityUtils.isInvisible(entity)) {
 				if (entity instanceof Player player) {
 					Owner viewingPlayer = new Owner(player);
 
@@ -98,9 +100,12 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 				else if (activatedOnlyByPlayer())
 					return false;
 
-				level.setBlockAndUpdate(worldPosition, state.setValue(RetinalScannerBlock.POWERED, true));
+				level.setBlockAndUpdate(worldPosition, state.cycle(RetinalScannerBlock.POWERED));
 				BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.RETINAL_SCANNER.get());
-				level.scheduleTick(new BlockPos(worldPosition), SCContent.RETINAL_SCANNER.get(), getSignalLength());
+
+				if (signalLength > 0)
+					level.scheduleTick(new BlockPos(worldPosition), SCContent.RETINAL_SCANNER.get(), signalLength);
+
 				return true;
 			}
 		}
@@ -117,6 +122,14 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 		}
 
 		return false;
+	}
+
+	@Override
+	public void onOptionChanged(Option<?> option) {
+		if (option.getName().equals(signalLength.getName())) {
+			level.setBlockAndUpdate(worldPosition, getBlockState().setValue(RetinalScannerBlock.POWERED, false));
+			BlockUtils.updateIndirectNeighbors(level, worldPosition, getBlockState().getBlock());
+		}
 	}
 
 	@Override