+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockChangeDetectorBlock.java
@@ -77,8 +77,10 @@ public class BlockChangeDetectorBlock extends DisguisableBlock {
 
 	@Override
 	public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
-		level.setBlockAndUpdate(pos, state.setValue(POWERED, false));
-		BlockUtils.updateIndirectNeighbors(level, pos, this);
+		if (state.getValue(POWERED)) {
+			level.setBlockAndUpdate(pos, state.setValue(POWERED, false));
+			BlockUtils.updateIndirectNeighbors(level, pos, this);
+		}
 	}
 
 	@Override