+++ src/main/java/net/geforcemods/securitycraft/blocks/InventoryScannerBlock.java
@@ -239,7 +239,7 @@ public class InventoryScannerBlock extends DisguisableBlock {
 		if (!(level.getBlockEntity(pos) instanceof InventoryScannerBlockEntity be))
 			return 0;
 
-		return be.isModuleEnabled(ModuleType.REDSTONE) && be.isProvidingPower() ? 15 : 0;
+		return be.isProvidingPower() ? 15 : 0;
 	}
 
 	@Override
@@ -282,7 +282,7 @@ public class InventoryScannerBlock extends DisguisableBlock {
 
 		@Override
 		public boolean isPowering(Level level, BlockPos pos, BlockState state, BlockEntity be, Direction direction, int distance) {
-			return ((InventoryScannerBlockEntity) be).isModuleEnabled(ModuleType.REDSTONE) && ((InventoryScannerBlockEntity) be).isProvidingPower();
+			return ((InventoryScannerBlockEntity) be).isProvidingPower();
 		}
 
 		@Override